requirements.txt
google-api-python-client
json
feedparser
python-dotenv